Aplicacion de autenticacion, tests y middleware personalizado

+++ app/Http/Controllers/personaController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use App\Models\persona;
 
 class personaController extends Controller
 {
@@ -13,6 +14,7 @@ class personaController extends Controller
     
     //Metodo para almacenar registros en la tabla
     public function store(Request $request){
+        // Validación de los datos de entrada
         $request->validate([
             'nombres' => 'required|string|max:50',
             'apellidos' => 'required|string|max:50',
@@ -21,13 +23,21 @@ class personaController extends Controller
             'direccion' => 'nullable|string',
             'telefono' => 'nullable|string|max:15'
         ]);
-        return persona::create($request->all());
+        $personas=persona::create($request->all());
+        return response()->json([
+            'mensaje' => 'Persona creada correctamente.',
+            'persona' => $personas
+        ], 201);
     }
 
     //Metodo para mostrar persona por ID
     public function show($id)
     {
-        $persona = Persona::findOrFail($id);
+        $persona = Persona::find($id);
+
+        if (!$persona) {
+            return response()->json(['mensaje' => 'Persona no encontrada.'], 420);
+        }
         return $persona;
     }
 
@@ -35,8 +45,14 @@ class personaController extends Controller
     //Metodo para actualizar una persona
     public function update(Request $request, $id)
     {
-        $persona = Persona::findOrFail($id);
+        $persona = persona::findO($id);
 
+
+        if (!$persona) {
+            return response()->json(['mensaje' => 'Persona no encontrada.'], 404);
+        }
+
+        // Validación de los datos de entrada
         $request->validate([
             'nombres' => 'sometimes|required|string|max:50',
             'apellidos' => 'sometimes|required|string|max:50',
@@ -47,16 +63,24 @@ class personaController extends Controller
         ]);
 
         $persona->update($request->all());
-        return $persona;
+        return response()->json([
+            'mensaje' => 'Persona actualizada correctamente.',
+            'persona' => $persona
+        ]);
     }
 
     //Metodo para eliminar una persona
     public function destroy($id)
     {
-        $persona = Persona::findOrFail($id);
+        $persona = Persona::find($id);
+
+        if (!$persona) {
+            return response()->json(['mensaje' => 'Persona no encontrada.'], 404);
+        }
+
         $persona->delete();
 
-        return response()->json(['mensaje' => 'Persona eliminada correctamente.']);
+        return response()->json(['mensaje' => 'Persona eliminada correctamente.'], 200);
     }
 
 }
